+++ src/WebJobs.Extensions.DurableTask/Listener/TaskCommonShim.cs
@@ -12,6 +12,8 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
     /// </summary>
     internal abstract class TaskCommonShim : TaskOrchestration
     {
+        private readonly TaskCompletionSource<Exception> timeoutTaskCompletionSource = new TaskCompletionSource<Exception>();
+
         public TaskCommonShim(DurableTaskExtension config)
         {
             this.Config = config ?? throw new ArgumentNullException(nameof(config));
@@ -22,6 +24,13 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
         internal DurableTaskExtension Config { get; private set; }
 
         protected Func<Task> FunctionInvocationCallback { get; private set; }
+
+        internal Task<Exception> TimeoutTask => this.timeoutTaskCompletionSource.Task;
+
+        internal void TimeoutTriggered(Exception exception)
+        {
+            this.timeoutTaskCompletionSource.TrySetResult(exception);
+        }
 
         public void SetFunctionInvocationCallback(Func<Task> callback)
         {