+++ release_notes.md
@@ -3,6 +3,7 @@
 - Added a boolean property `IncludeDeleted` to `EntityQuery` which controls whether to return deleted entities.
 
 ## Bug fixes
+- Fixed handling of function timeouts inside entity and activity functions and added tests
 
 ## Breaking Changes